renamed file names to be compatible with Django
