+++ loaddata.sql
@@ -85,6 +85,6 @@ CREATE TABLE "Categories" (
   "label" varchar
 );
 
-INSERT INTO Categories ('label') VALUES ('News');
-INSERT INTO Tags ('label') VALUES ('JavaScript');
+INSERT INTO rareapi_category ('label') VALUES ('News');
+INSERT INTO rareapi_tag ('label') VALUES ('JavaScript');
 INSERT INTO Reactions ('label', 'image_url') VALUES ('happy', 'https://pngtree.com/so/happy');
