+++ rareapi/models/__init__.py
@@ -1,5 +1,5 @@
-from .RareUser import RareUser
-from .Post import Post
-from .Category import Category
-from .PostTag import PostTag
-from .Tag import Tag
+from .rareUser import RareUser
+from .post import Post
+from .category import Category
+from .postTag import PostTag
+from .tag import Tag
