+++ rareapi/views/__init__.py
@@ -1,4 +1,4 @@
-from .Posts import PostViewSet
-from .Categories import CategoryViewSet
-from .Tags import TagViewSet
+from .posts import PostViewSet
+from .categories import CategoryViewSet
+from .tags import TagViewSet
 from .users import UserViewSet
