+++ rareapi/views/categories.py
